+++ src/ollist.c
@@ -36,6 +36,7 @@ btr_ollist_s BTR_OLList_make(
 }
 btr_ollist_s BTR_OLList_clone(const btr_ollist_s *list, btr_allocator_s *allocator)
 {
+    BTR_panicIf(!list, "`list` is null");
     btr_ollist_s newList = BTR_OLList_make(list->itemSize, allocator);
     BTR_OLLIST_FOREACH(list, i)
         memcpy(BTR_OLList_append(&newList), i, list->itemSize);
@@ -43,24 +44,28 @@ btr_ollist_s BTR_OLList_clone(const btr_ollist_s *list, btr_allocator_s *allocat
 }
 void *BTR_OLList_append(btr_ollist_s *this)
 {
+    BTR_panicIf(!this, "`this` is null");
     void *data = allocateData(this);
     BTR_BLList_append(&this->underlying, data);
     return data;
 }
 void *BTR_OLList_prepend(btr_ollist_s *this)
 {
+    BTR_panicIf(!this, "`this` is null");
     void *data = allocateData(this);
     BTR_BLList_prepend(&this->underlying, data);
     return data;
 }
 void *BTR_OLList_insert(btr_ollist_s *this, long index)
 {
+    BTR_panicIf(!this, "`this` is null");
     void *data = allocateData(this);
     BTR_BLList_insert(&this->underlying, data, index);
     return data;
 }
 void BTR_OLList_pop(btr_ollist_s *this, long index, void *buffer)
 {
+    BTR_panicIf(!this, "`this` is null");
     void *popped = BTR_unwrap(BTR_BLList_pop(&this->underlying, index));
     if (buffer)
         memcpy(buffer, popped, this->itemSize);
@@ -68,41 +73,50 @@ void BTR_OLList_pop(btr_ollist_s *this, long index, void *buffer)
 }
 btr_container_ptr_r BTR_OLList_get(const btr_ollist_s *this, long index)
 {
+    BTR_panicIf(!this, "`this` is null");
     return BTR_BLList_get(&this->underlying, index);
 }
 btr_container_ptr_r BTR_OLList_first(const btr_ollist_s *this)
 {
+    BTR_panicIf(!this, "`this` is null");
     return BTR_BLList_first(&this->underlying);
 }
 btr_container_ptr_r BTR_OLList_last(const btr_ollist_s *this)
 {
+    BTR_panicIf(!this, "`this` is null");
     return BTR_BLList_last(&this->underlying);
 }
 btr_container_idx_r BTR_OLList_indexOf
     (btr_ollist_s *list, void *value, bool (*cmp)(const void *, const void *))
 {
+    BTR_panicIf(!list, "`list` is null");
     return BTR_BLList_indexOf(&list->underlying, value, cmp);
 }
 size_t BTR_OLList_len(const btr_ollist_s *this)
 {
+    BTR_panicIf(!this, "`this` is null");
     return BTR_BLList_len(&this->underlying);
 }
 bool BTR_OLList_isEmpty(const btr_ollist_s *this)
 {
+    BTR_panicIf(!this, "`this` is null");
     return BTR_BLList_isEmpty(&this->underlying);
 }
 void BTR_OLList_reverse(btr_ollist_s *this)
 {
+    BTR_panicIf(!this, "`this` is null");
     BTR_BLList_reverse(&this->underlying);
 }
 void BTR_OLList_free(btr_ollist_s *this)
 {
+    BTR_panicIf(!this, "`this` is null");
     BTR_BLLIST_FOREACH(&this->underlying, i)
         BTR_Allocator_deallocate(this->allocator, i);
     BTR_BLList_free(&this->underlying);
 }
 void BTR_OLList_clear(btr_ollist_s *this)
 {
+    BTR_panicIf(!this, "`this` is null");
     BTR_BLLIST_FOREACH(&this->underlying, i)
         BTR_Allocator_deallocate(this->allocator, i);
     BTR_BLList_clear(&this->underlying);